+++ test/fixtures/frameworks/bar/src/config/config.default.ts
