+++ test/fixtures/frameworks/foo/src/config/config.default.ts
